backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "tescord-backend"
version = "0.1.0"
description = "FastAPI backend for Altgramm"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0",
  "fastapi>=0.116.0",
  "pywebpush>=2.0.3",
  "python-multipart>=0.0.20",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.43",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
